refactor: new lib structure with internals
