+++ projects/spectator/jest/src/public_api.ts
@@ -1,9 +1,8 @@
 /// <reference types="jest" />
 /// <reference path="./matchers-types.d.ts" />
 export * from './dom-selectors';
-export * from './host';
+export * from './spectator-with-host';
 export * from './http';
-export * from './internals';
 export * from './mock';
 export * from './service';
 export * from './spectator';
