+++ projects/spectator/jest/src/spectator.ts
@@ -1,8 +1,7 @@
-import { createTestComponentFactory as baseCreateTestComponentFactory, isType, SpectatorOptions } from '@netbasal/spectator';
+import { createTestComponentFactory as baseCreateTestComponentFactory, Spectator as BaseSpectator, SpectatorOptions, Token } from '@netbasal/spectator';
 import { Type } from '@angular/core';
-import { Spectator } from './internals';
-import { CreateComponentOptions } from '../../src/lib/types';
-import { mockProvider } from './mock';
+import { CreateComponentOptions, isType } from '../../src/lib/types';
+import { mockProvider, SpyObject } from './mock';
 
 export function createTestComponentFactory<Component>(typeOrOptions: SpectatorOptions<Component> | Type<Component>): (options?: CreateComponentOptions<Component>) => Spectator<Component> {
   return baseCreateTestComponentFactory(
@@ -16,4 +15,10 @@ export function createTestComponentFactory<Component>(typeOrOptions: SpectatorOp
           ...typeOrOptions
         }
   );
+}
+
+export class Spectator<C> extends BaseSpectator<C> {
+  get<T>(type: Token<T> | Token<any>, fromComponentInjector = false): SpyObject<T> {
+    return super.get(type, fromComponentInjector) as SpyObject<T>;
+  }
 }
