+++ projects/spectator/src/lib/base/base-module-metadata.ts
@@ -1,0 +1,6 @@
+import { TestModuleMetadata } from '@angular/core/testing';
+import { Type } from '@angular/core';
+
+export interface BaseModuleMetadata extends TestModuleMetadata {
+  entryComponents?: Type<any>[];
+}
