+++ projects/spectator/src/lib/index.ts
@@ -1,17 +1,17 @@
-export * from './config';
-export * from './dispatch-events';
+export * from './spectator/spectator';
+export * from './spectator/options';
+export * from './spectator/create-factory';
+export * from './spectator/initial-module';
+export * from './spectator/host-component';
+export * from './spectator/module-metadata';
+
+export * from './spectator-with-host/spectator-with-host';
+export * from './spectator-with-host/create-factory';
+
 export * from './dom-selectors';
-export * from './element-focus';
-export * from './event-objects';
-export * from './host';
 export * from './http';
-export * from './internals';
-export * from './types';
 export * from './matchers';
 export * from './mock';
-export * from './rgb-to-hex';
 export * from './service';
-export * from './spectator';
 export * from './token';
-export * from './type-in-element';
-export * from './globals-injections';
+export * from './types';
