+++ projects/spectator/src/lib/internals/dispatch-events.ts
