+++ projects/spectator/src/lib/internals/element-focus.ts
