+++ projects/spectator/src/lib/internals/event-objects.ts
