+++ projects/spectator/src/lib/internals/rgb-to-hex.ts
