+++ projects/spectator/src/lib/matchers.ts
@@ -2,7 +2,7 @@
 declare const require: Function;
 
 const $ = require('jquery');
-import { hex2rgb, isHex, trim } from './rgb-to-hex';
+import { hex2rgb, isHex, trim } from './internals/rgb-to-hex';
 
 const hasProperty = (actual, expected) => {
   return expected === undefined ? actual !== undefined : actual === expected;
