+++ projects/spectator/src/lib/spectator/host-component.ts
@@ -1,0 +1,6 @@
+import { Component } from '@angular/core';
+
+@Component({
+  template: ''
+})
+export class HostComponent {}
